LPD-59495 Move Breadcrumb, SpaceSticker and BackButtonManagementBar components to common folder
