+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/FDSPropsTransformer/cell_renderers/SpaceRenderer.tsx
@@ -5,7 +5,7 @@
 
 import React from 'react';
 
-import SpaceSticker from '../../components/SpaceSticker';
+import SpaceSticker from '../../../common/components/SpaceSticker';
 
 const SpaceRenderer = ({itemData, value}: {itemData: any; value: string}) => {
 	return (
