+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/categorization/components/CategorizationSpaces.tsx
@@ -10,9 +10,9 @@ import ClayMultiSelect from '@clayui/multi-select';
 import {sub} from 'frontend-js-web';
 import React, {ChangeEvent, useEffect, useState} from 'react';
 
+import SpaceSticker from '../../../common/components/SpaceSticker';
 import SpaceService from '../../../common/services/SpaceService';
 import {LogoColor} from '../../../common/types/Space';
-import SpaceSticker from '../../components/SpaceSticker';
 
 type Space = {
 	displayType?: string;
