+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/main/components/SpacesDisplay.tsx
@@ -7,8 +7,8 @@ import Badge from '@clayui/badge';
 import {ClayTooltipProvider} from '@clayui/tooltip';
 import React from 'react';
 
+import SpaceSticker from '../../common/components/SpaceSticker';
 import {Space} from '../../common/types/Space';
-import SpaceSticker from './SpaceSticker';
 
 export interface SpaceDisplayProps {
 	spaces: Space[];
